backend/requirements.txt
# requirements.txt - 后端 Python 依赖列表

# Flask 核心依赖
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# 数据库依赖
pymongo==4.5.0

# 文档解析依赖
PyMuPDF==1.23.9        # PDF 解析 (fitz)
python-docx==1.1.0     # DOCX 解析
